test/pylib: move environment preparation into runner and drop suite.base
Move prepare_dir/prepare_dirs, the 3rd-party service startup
(start_3rd_party_services, _make_service_logger) and the
PYTEST_TESTS_LOGS_FOLDER constant from test/pylib/suite/base.py into
test/pylib/runner.py. With nothing left in suite/base.py, remove the
file.
